refactor: test_info.py code moved to test_lit_server.py
